Enable concurrent adds to IdNamesPJContainer.
The zope.container behaviour is to check before inserting if the key
already exists in a container.  If so, if the existing value is the
same, nothing happens, but if it's different, a KeyError is raised.

On pjpersist, such a check creates a SELECT query before the INSERT
query.  With different keys, multiple INSERTs would succeed, but the
SELECT makes the transactions conflict on each other.

So, in this commit, we skip the check if the key is in the container if
the key is a generated id.

With IdNamesPJContainer, usually the name is generated and presumed to
be unique, so we can be sure that the object with a given key does not
exist.  In the unlikely case that the just-created ids clash, a
conflict would prevent any data loss.